Add: test_no_connection
