+++ test/webserv/unit/http/test_case/test_case.hpp
@@ -1,0 +1,19 @@
+#ifndef TEST_CASE_HPP_
+#define TEST_CASE_HPP_
+
+#include <list>
+
+namespace server {
+
+class VirtualServer;
+typedef std::list<const VirtualServer *> VirtualServerAddrList;
+
+} // namespace server
+
+namespace test {
+
+int test_no_connection(const server::VirtualServerAddrList &server_infos);
+
+}
+
+#endif
